feat: Add inline keyboard options to Telegram message response
This commit modifies the Telegram webhook route to include inline keyboard options in the message response. Now, when the user sends a message and the username is "luisignaciocc", the bot will reply with a message and display two buttons: "Aceptar" and "Cancelar". This provides the user with the option to accept or cancel the message.